renaming spatial.py to subdomain.py
