+++ src/osyris/spatial/subdomain.py
